+++ deepxube/trainers/train_heur.py
@@ -4,9 +4,7 @@ from deepxube.base.heuristic import HeurNNet
 from deepxube.base.updater import UpdateHeur
 from deepxube.base.trainer import Train, update_optimizer
 from deepxube.utils.timing_utils import Times
-from deepxube.trainers.utils.train_utils import train_heur_nnet_step, ctgs_summary
-
-import torch.nn as nn
+from deepxube.trainers.utils.train_utils import train_nnet_step, ctgs_summary
 
 import numpy as np
 from numpy.typing import NDArray
@@ -20,18 +18,17 @@ class TrainHeur(Train[HeurNNet, UpdateHeur]):
 
     def _train_itr(self, batch: List[NDArray], first_itr_in_update: bool, times: Times) -> float:
         start_time = time.time()
-        inputs_batch_np: List[NDArray] = batch[:-1]
         ctgs_batch_np: NDArray = batch[-1]
         ctgs_batch_np = np.expand_dims(ctgs_batch_np.astype(np.float32), 1)
+        batch = batch[:-1] + [ctgs_batch_np]
 
         self.nnet.train()
         update_optimizer(self.optimizer, self.nnet, self.status.itr)
-        ctgs_batch_nnet, loss = train_heur_nnet_step(self.nnet, inputs_batch_np, ctgs_batch_np, self.optimizer, nn.MSELoss(), self.device,
-                                                     self.status.itr, self.train_args, self.train_start_time)
+        fwd_tr_tensors, loss = train_nnet_step(self.nnet, batch, self.optimizer, self.device, self.status.itr, self.train_args, self.train_start_time)
         self.writer.add_scalar("train/loss", loss, self.status.itr)
 
         if first_itr_in_update:
-            self.train_summary.itr_to_in_out[self.status.itr] = (ctgs_batch_np, ctgs_batch_nnet)
+            self.train_summary.itr_to_in_out[self.status.itr] = (ctgs_batch_np, fwd_tr_tensors[0].cpu().data.numpy())
         times.record_time("train", time.time() - start_time)
         return loss